Added ReceiverAttribute and updated actor handling methods
This commit introduces several changes to improve the handling of actors and their methods. A new `ReceiverAttribute` class has been added, which can be applied to methods.

`ActorGenerator.cs` has been updated to generate receiver methods for each entry node in the actor with the `Receiver` attribute. The `GenerateResultReceivers` method has been renamed to `GenerateResultAcceptors` and updated to generate acceptor methods. A new `GenerateReceiverMethod` method has been added to generate the code for the receiver methods.